+++ src/ol_openedx_events_handler/CHANGELOG.rst
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+Version 0.2.0 (2026-04-17)
+---------------------------
+
+* Added LMS receiver for ``COURSE_GRADE_NOW_PASSED`` to trigger certificate
+  creation callbacks in MIT systems.
+
 Version 0.1.0 (2026-03-17)
 ---------------------------